+++ server/prisma/migrations/20260916162026_agenda_evento_tipo_otro/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "TipoEventoAgenda" ADD VALUE 'OTRO';